// generated/src/app/api-models/cap-adjustments-work.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapAdjustmentsWork {
  seqCcalcId: number;
  seqCapAdjustmentWork: number;
  capModelId: string;
  firstContractMonth: Date;
  capEntityCode: string;
  capPoolId: string;
  adjustmentDate: Date;
  seqProvId: number;
  seqVendId: number;
  capMonthRun: Date;
  description: string;
  applyTo: string;
  capStoplossId: string;
  adjustmentAmt: number;
  status: string;
  companyCode: string;
  glRefCode: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqMembrId: number;
  seqVendAddress: number;
  adjustReason: string;
  userDefine1: string;
  userDefine2: string;
  userDefine3: string;
  userDefineDate: Date;
  postCode: string;
  capEntityId: string;
}